Trading-fin/tradingfin3/Trading-patched/Trading/Trading-fin/requirements.txt
pandas
numpy
requests
openai
python-dotenv
pandas-ta
ta
joblib
xgboost
lightgbm
transformers
torch
passlib
cryptography
fastapi
uvicorn
scikit-learn
python-telegram-bot
aiohttp
prometheus_client
python-jose
python-multipart
cloudscraper
psutil
